+++ TinkoffASDKCore/TinkoffASDKCore/API/Models/PaymentFinishResponseStatus.swift
@@ -1,0 +1,32 @@
+//
+//
+//  PaymentFinishResponseStatus.swift
+//
+//  Copyright (c) 2021 Tinkoff Bank
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+
+import Foundation
+
+public enum PaymentFinishResponseStatus {
+    /// Требуется подтверждение 3DS v1.0
+    case needConfirmation3DS(Confirmation3DSData)
+
+    /// Требуется подтверждение 3DS v2.0
+    case needConfirmation3DSACS(Confirmation3DSDataACS)
+
+    /// Успешная оплата
+    case done
+}
